requirements.txt
boto3
pandas
redis
torch
torchaudio
numpy
scipy
scikit-learn
tqdm
scenedetect
opencv-python
python_speech_features
torchvision
ffmpeg-python
gdown
youtube-dl
facenet-pytorch
Pillow
python-dotenv
av
torchvision
mediapipe
